kafka/requirements.txt
kafka-python>=2.0.2
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sentence-transformers>=2.2.2
pymongo>=4.4.0
python-dotenv>=1.0.0
tqdm>=4.65.0